Implement environment variable to disable ESLint during storefront builds.
Inside the administration webpack.base.config.js, the option to disable linting already exists,
so it seems sensible to provide this for storefront as well.
This is helpful, once you use a custom eslint configuration inside a plugin that is not compatible with
the shopware linting process (due to missing plugins, ...).