fix: add missing custom fields to order AddressTransformer
Since the order address is a snapshot of the customer address, all custom fields that may be added to the respective customer address, should be cloned too.
With this change it's also possible to access the custom fields of the customer address when querying the order or one of its addresses via api.