Fix link to product from order line item
The administration's client-side route "sw.product.detail" expects a
supplied parameter to be a product id.

This patch adjusts the link generation to use the known product id from
the order line item record.

The product id is already properly used to generate the link in the
context menu (instead of using the order line item "identifier").

While the previously used "identifier" from the order line item may
match the product id, it may also be different.

Order line items created by the LineItemTransformer, given cart line
items originally created by the CartLineItemController using the
ProductLineItemFactory, will have such an "identifier" matching the
product id.

However, that need not apply to "identifiers" of order line items
created elsewhere. The OrderConverter from the SwagMigrationAssistant
plugin is such a counterexample, as it sets the "identifier" to the
order line item's id (not the product's id).