Fix parse boolean value error with elasticsearch
If I assign a dynamic product group to a category and assign the filter "active = 1" or a other boolean filter to the dynamic product group, an error "failed to create query: Can't parse boolean value [1], expected [true] or [false]" occures on searching. A fallback to the normal database search will happen - it shouldn't.