Add additional --sales-channel-doman-url option to "theme:dump" command
For Stores with more than one storefront sales-channel installed, we want to be able to specify the theme we want to dump. Doing so using the theme ID is tedious for the developer, since it requires additional lookup and or documentation steps. In this non-breaking pull request I propose the addition of an input option that allows us to specify the theme by the sales-channel domain url instead. This pull request is an important stepping stone to make other development tasks, like `./psh.phar storefront:hot` and `./psh.phar storefront:hot-proxy` usable for development environments that work on more that sales-channel theme at the same time.