(administration) Declare data property in toOne Associations, even if the value is NULL
If we associate entity extensions in the administration, and the association is a toOne association, we currently
do NOT declare the property in the data object, if the joined entity is missing (the value is NULL). I argue, that
the developer expects the property to be declared in the data object, even if no joined entity exists. In
just cases the data property should be declared with the value NULL.