Fix error in certain products w/o variants
After migration from Shopware 5, this error occurs on several products which don't have variants assigned. The check $product->getOptions() always returns an array, so the check can never be null, which leads to the following error down the line:

count(): Parameter must be an array or an object that implements Countable

The right check is to check $product->getOptionIds() against null, which in fact does happen.