(bugfix) Fix DBAL Entity Hydration when one entity is hydrated multiple times with different associations loaded
It is possible that one entity is being hydrated multiple times with differently loaded associations. In these cases the current caching strategy fails to hydrate the entities correctly, because it caches the hydrated entities ONLY by its unique identifier, DISREGARDING the loaded associations of the entity. In situations, where we hydrate the same entity (same ID) two ore more times with the different associations, the entity of the first hydration is REUSED. This can lead to missing association data in the loaded entity.

** Use Case **

Think of the following situation
PickupPoint
|- id
|- warehouseId
|- warehouse
|- |- id
|- |- zipcodeId
|- |- zipcode
|- |- |- id <--- ID 123
|- |- |- countryId
|- zipcodeId
|- zipcode
|- |- id <--- ID 123
|- |- countryId
|- |- country
|- |- |- id
|- |- |- iso

In the use case above the pickup point and the warehouse share the same zipcode (which in this use case is an entity, not a string). While the Search Query DOES NOT LOAD the country association of the zipcode entity associated with the warehouse, it does load the associated country of the zipcode entity associated with the pickup point.

The above use case is currently not hydrated correctly, since the first time the zipcode is being hydrated, the country data is missing, the country data is also missing for the second zipcode entity, since it is being reused, using the current caching