Make sure one to one entities are encoded correctly by the JsonApiEncoder
If one of our entities (Entity A) has a one to one association with a different entity (Entity B), it is very likely that
this entity has a relationship back to the first entity.

Entity A < ------ > Entity B

Searching for Entity A, while also loading Entity B ($criteria->addAssociation('entityB') would result in a result like this

$entiyA
|-- $entiyB
|-- |-- $entityA
|-- |-- |-- $entityB = null

Or, if $entityB is implemented via an entity extension:

$entityA
|-- $extensions
|-- |-- $entityB
|-- |-- |-- $entityA
|-- |-- |-- |-- $extensions
|-- |-- |-- |-- |-- $entiyB = null

The problem we have in these situations is the following, entity A is serialized first in the process of serializing entityB:

Process:
Serialize Entity A --> Serialize Relationships / Extensions of Entity A --> Serialize Entity B (which is related / an extension of Entity A) --> Serialize relationships / Extensions of Entity B --> Serialize Entity A (which is a related / an extension of Entity B)

This process naturally ends at some point, as we are walking down the tree. HOWEVER, we do have a problem once we add the serialized data to the included section of the record.
Since we include All relationships / extensions of Entity A the first time in the process of serializing Entity B, we do not add the data again when we come back to Entity A when we walk back up the tree. This results in the fact that the data value of Entity B remains NULL in the included part, even though the association has been loaded in the search query.

This phenomenon is easier understood running the unit test, which i provided with this commit. Without the fix, the toOne Relationship is not correctly serialized.