Don't limit the number of loaded state transitions
The states loaded via the getAllStates() method are used to determine the available state transitions that an order can go through. By not explicitly setting the limit of the criteria object to NULL, the criteria receives the default LIMIT value of 25. Thus, only the first 25 state transitions are loaded from the database.