Fix category.cmsPage.locked error
With Shopware 6.4.5.0 a new bug was introduced that can cause the application to fail, while saving category detail pages. 
If a category has no cmsPage assigned to it, the check if that cmsPage is locked or not will cause a javascript error and results in the user waiting for the saving process to finish. We therefore need to check if the category has an cmsPage assigned to it first.