Better compatibility of admin with browser extensions
There are some browser extensions like Bitwarden which insert a diff on top of the body. This causes a very annoying shift in the menu, making navigation items unreachable when the notification is displayed.

This commit fixes this by checking the bounding rect of the #app container (which in standard cases is 0)