Search term interpreter tuning.
A good search result was getting a low score using levenshtein function. E.g., the needle is DW0DW11457 and the haystack is DW0DW11457-0XU. This would get it a score of 4 which is too high to give it any advantage. Simple check if the given needle is in a haystack fix this.