fix: ensure sequential indexes for configuration arrays
When a plugin defines a config.xml with elements using <flag> these elements are removed if the feature flag is not enabled. This could lead to a JavaScript error when opening the Plugin configuration in the Shopware Administration.

It happens because if a configuration element is removed the resulting PHP array is non-sequentially indexed and when using json_encode this results in an Object instead of an Array. The Administration then tries to call JavaScript Array methods on an Object which throws an Error.