NEXT-32108 - gitlab CI - MySQL/MariaDB normalization + mariadb pre-release testing
The overriding of the entrypoint of mysql and mariadb was a little moot in they both accepted command line options and didn't need mysqld/mariadbd there.

Normalizes the command of special options using the --loose-* prefix that both MySQL and MariaDB will ignore if not recognised for uniform. Added the InnoDB force primary key as equivalent.

per #3273 added quay.io/mariadb-foundation/mariadb-devel:verylatest so any regressions can be found pre-release.