Update action.html.twig to include css class for detail button
As a developer, I want to be able to react to clicks on the detail button (in my case, to track "select_item" events for Google Analytics. Right now, I would have to target all buttons with "btn-light" class and I want it to be more specific, so I added a new class "btn-detail", similar to "btn-buy".