Do not alter distPath for symlinked plugins
When plugins are installed via path repository and therefore symlinked,
they will not get detected as being installed in `/vendor` due to usage
of `realpath()` when the base path is determined.

This corrects the behaviour to only change the path
from `plugin-base/src/Resources`
to   `plugin-base/src/<TechnicalName>/Resources`
when the first one does not exist.