feat: configuration for meta tag "robots"
Add configuration option at Settings > Basic information to define the
default content of meta tag "robots".

Using the hard coded value as a fallback on a null-ish value, because
updating an existing project will not auto-load the defaultValue into
DB and people may save an empty value, thinking it would fall back to
the default/previous behaviour.

Initially, I thought to add the configuration option to "Settings >
SEO", but that configuration does not support saving settings for an
individual sales channel.