fix: Consider only plugins located in custom directory
If a plugin no longer executes the composer commands, it should only be removed via composer during update, if it is located in the "custom/(static-)plugins" directory.
If a plugin was directly required via composer and lives only in the "vendor" directory, it should not be considered, even if the "executeComposerCommands" flag is false.