feat: native en-US in installer and language based pre-selections
en-US is treated as a first-party language by extending tests to en-US
and ensuring that not only de always works when selected as a language
but en-US does, too.

Currencies and countries are now preselected when changing the language
in the configuration, providing a better overall experience for
customers during the first time installation.

Moved the language selection from the footer into a header select field
for a more polished look-and-feel and to have a clean solution to
provide more languages in the near future.

Improved contrast on the versio number and the label to enhance
readability and accessibility.