fix: backwards compatibility to Shopware 6.6 class
In Shopware 6.6 `MediaThumbnailEntity` objects were allowed to have null
for members that were not strictly typed but the getters/setters
implicated that null would not be allowed.

As these objects are stored serialized in the database, when cart
compression is enabled, it's impossible to `unserialize()` an object
again, as the strict type now restricts `null` as a value.

With this change `$mediaId` will behave the same as in 6.6 by allowing
null, but implicitly requiring a string.

Fixes #10040