feat: use flysystem for downloaded languages
The original implementation used native functions to download
translation files into the `src` directory. With this change, the
translation files will be downloaded to the private filesystem which can
be configured to be on a remote system or anywhere League's Flysystem
allows.

The snippet handling now supports loading snippets from different
sources. Paths are passed via a collection, allowing to add flags like
the target filesystem to a path.