feat: dispatch event before running aggregations
Dispatch an event before running aggregations when entities are
searched. This allows listeners to sanitize the aggregations before
letting them run on the database.

This is a WIP POC. I intend to remove the medie specific names and
namespaces and put the functionality in a modular service so other
listeners can make use of this, too. For easier discussion and to have a
working fix, I implemented this in the media namespace as that's the
place currently having issues.