feat: Allow IDN for email addresses
The backend form handling already allows IDN for emails, but the
validator did not allow that.

With this change, the validator checks for IDN, uses the browser's
native URL method to convert to punycode any runs the same regular
expression on the punycode email address like an regular ASCII
addresses.

This is part of the solution to
https://github.com/shopware/shopware/issues/5356