feat: Always create base snippet set
When a language/translation/snippet set is installed using the core
functionality, always create a snippet set representing that
translation. Don't rely on snippet sets with the same locale created by
other plugins or the user.

This fixes issues when, e.g. a snippet set exists from the installation
of the Language Pack but the Language Pack is uninstalled and removes
its data. Without a base snippet set, the snippet files would still be
available, but not selectable for a sales channel